extracted product-context to new file
